Update the about section
This update makes clear that we're not cinelerra.org. It also tells people why we're not in touch with them, without being emotional.

I've removed the sections about collaborative development and what the website offers. The dev stuff is essentiall being told in main_repo.php. What this website offers can be seen (or put on the frontpage later).

Removed the video by Andraz because it is broken.
Added corresponding mailing list messages and fixed the broken ones.