about.php: Clear up situation with merging back HV
In our recent reddit thread someone asked this:

"Does Cinelerra-HV ever merge Cinelerra-CV own fixes back too, or is it a one-way "collaboration?"
This is unclear too and if it does not happen, it could be a problem, because software would be slowly diverging every year.
That means that if they want to continue merging HV code, they need to always keep the same code structure for instance,
forbidding themselves from doing any deeper code change if they deemed fit."

Deeper code changes and enhancements to CV are not forbidden just for the sake of HV merges, cleared that up.

Also told people that we've fallen a bit behind with merging HV, but it is still our goal to do it.